refactor: extract tool accumulator and safety mixin from litellm driver and security service
